requirements/requirements.txt
pandas
requests
pyspark==3.5.0
psycopg2-binary
apache-airflow-providers-postgres